add cs

+++ cpp/08_parallelism/2_mkl/3_vector-multiplication-dot-product/mkl-dot-product.cpp
@@ -24,14 +24,14 @@ int main(int argc, char *argv[]) {
     exp = 5;
   cout << "exp,vector_size,result,takes(ms)\n";
   for (int e = 0; e < exp; ++e) {
-    long arr_size = pow(base, e);
+    size_t arr_size = pow(base, e);
     cout << fixed << setw(3) << e << ", " << setw(15) << arr_size << ", ";
     vector<double> vec_a;
     vector<double> vec_b;
-    for (long i = 0; i < arr_size; ++i) {
+    for (size_t i = 0; i < arr_size; ++i) {
       vec_a.push_back(get_random_0_to_1());
     }
-    for (long i = 0; i < arr_size; ++i) {
+    for (size_t i = 0; i < arr_size; ++i) {
       vec_b.push_back(get_random_0_to_1());
     }
